Conflicts:
	src/main/java/com/jorgerubira/ejerciciosjava/Ejercicio00PruebaDeTestBasica.java
	src/main/java/com/jorgerubira/ejerciciosjava/Ejercicio01ConversionTipos.java
	src/main/java/com/jorgerubira/ejerciciosspringweb/services/Ejercicio03GestionAlumnosService.java
	src/main/resources/application.properties

+++ src/main/java/com/jorgerubira/ejerciciosjava/Ejercicio01ConversionTipos.java
@@ -3,23 +3,28 @@ package com.jorgerubira.ejerciciosjava;
 public class Ejercicio01ConversionTipos {
 
     //Devuelve la suma a+b;
-    public int ejemploSuma(int a, int b) {
-        return a + b;
+    public int ejemploSuma(int a, int b){
+       
+		return a+b;
     }
 
     //Convierte de String a int.
     public int textoAEntero(String valor) {
-        return Integer.parseInt(valor);
+       
+    	return  Integer.parseInt(valor);
+		
     }
 
     //Convierte de Float a int.
-    public int decimalesAEntero(Float valor) {
-        return valor.intValue();
+    public Float decimalesAEntero(Float valor){
+    	
+		return (float) valor.intValue();
+    	
     }
 
     //Convierte de int a Float.
-    public Float enteroAFloat(int valor) {
-        return (float) valor;
+    public Float enteroAFloat(int valor){
+    	return (float)valor ;
     }
 
     //Devuelve el siguiente carácter en el alfabeto (tabla ASCII).
